requirements.txt
docopt>=0.4.0
mock>=2.0.0
tox>=2.3.1
mkdocs>=0.15.3
python-bitcoinlib>=0.5.0
requests>=2.11.0
cert-schema>=1.2.18
merkle-proofs>=0.0.5
pyld>=0.7.1
enum34>=1.1.6
